backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.10.4
attrs==25.3.0
bcrypt==4.3.0
beautifulsoup4==4.12.3
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
ecdsa==0.19.1
exceptiongroup==1.2.2
fastapi==0.115.4
greenlet==3.2.0
h11==0.14.0
httpcore==1.0.8
httpx==0.27.2
idna==3.10
iniconfig==2.1.0
lxml==5.3.0
mysql-connector-python==9.1.0
outcome==1.3.0.post0
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
PySocks==1.7.1
pytest==8.3.5
pytest-asyncio==0.26.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
pytz==2025.2
requests==2.31.0
rsa==4.9.1
selenium==4.21.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.35
starlette==0.41.3
tomli==2.2.1
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzlocal==5.3.1
urllib3==2.4.0
uvicorn==0.32.0
wsproto==1.2.0